main: break out of search loop as soon as target elf section is found
There is no need in running over the rest of the elf section. In most
cases (and in this very particular example) the custom section is
appended in the end, so there is not much of a difference. Though, in
contrast to the previous implementation, if multiple sections with
matching names have been added, now the data is extracted from the first
one, while previously the data from the last one was used.

Signed-off-by: Johannes Zink <opensource@johannes-zink.de>